requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
pandas
Pillow
matplotlib
seaborn
scikit-learn
wandb
PyYAML
tqdm
mediapipe

# Jupyter Notebook & UI
jupyter
ipywidgets

# Testing
pytest
